pyproject.toml
[tool.poetry]
name = "collecting_module"
version = "1.0.0"
description = "api & crawling & etc"
authors = ["tmxkqotnl <tmxkqotnl@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.2"
requests = "^2.27.1"
lxml = "^4.9.0"
ipykernel = "^6.13.0"
beautifulsoup4 = "^4.11.1"
selenium = "^4.2.0"
black = "^22.3.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
